budget/requirements.txt
pydantic>=2.5
pyyaml>=6.0
pandas>=2.1
numpy>=1.26
openpyxl>=3.1
streamlit>=1.30
plotly>=5.18
pytest>=8.0
requests>=2.31
